+++ cli/lib/context/context.dart
@@ -6,6 +6,7 @@ import 'package:stax/external_command/external_command.dart';
 import 'package:stax/file/file_path_dir_on_uri.dart';
 import 'package:stax/git/git.dart';
 import 'package:stax/rebase/rebase_use_case.dart';
+import 'package:stax/settings/base_settings.dart';
 import 'package:stax/settings/repository_settings.dart';
 import 'package:stax/settings/settings.dart';
 
@@ -20,10 +21,10 @@ class Context {
   late final Settings settings = Settings();
   late final RepositorySettings? repositorySettings = RepositorySettings.load(
     this,
+    settings,
   );
+  late final BaseSettings effectiveSettings = repositorySettings ?? settings;
   late final RebaseUseCase? rebaseUseCase = RebaseUseCase.create(this);
-
-  RepositorySettings get assertRepositorySettings => repositorySettings!;
 
   RebaseUseCase get assertRebaseUseCase => rebaseUseCase!;