+++ cli/lib/settings/repository_settings.dart
@@ -2,10 +2,12 @@ import 'package:path/path.dart' as path;
 import 'package:stax/context/context.dart';
 import 'package:stax/context/context_git_get_repository_root.dart';
 import 'package:stax/settings/base_settings.dart';
+import 'package:stax/settings/key_value_store.dart';
+import 'package:stax/settings/settings.dart';
 
-class RepositorySettings extends BaseSettings {
-  static RepositorySettings? load(Context context) {
-    final root = context.getRepositoryRoot();
+class RepositorySettings extends KeyValueStore with BaseSettings {
+  static RepositorySettings? load(Context context, Settings settings) {
+    final root = context.withSilence(true).getRepositoryRoot();
     if (root == null) return null;
     return RepositorySettings(
       path.join(
@@ -15,8 +17,19 @@ class RepositorySettings extends BaseSettings {
         'stax',
         'settings.json',
       ),
+      settings,
     );
   }
 
-  RepositorySettings(super.path) : super.fromPath();
+  final Settings _settings;
+
+  RepositorySettings(super.path, this._settings) : super.fromPath();
+
+  @override
+  String get name => '--local';
+
+  @override
+  String? operator [](String key) {
+    return super[key] ?? _settings[key];
+  }
 }